+++ src/mate/src/App.php
@@ -11,15 +11,20 @@
 
 namespace Symfony\AI\Mate;
 
+use Mcp\Server\Transport\Stdio\RunnerControl;
+use Mcp\Server\Transport\Stdio\RunnerState;
 use Psr\Log\LoggerInterface;
 use Symfony\AI\Mate\Command\ClearCacheCommand;
+use Symfony\AI\Mate\Command\DebugCapabilitiesCommand;
+use Symfony\AI\Mate\Command\DebugExtensionsCommand;
 use Symfony\AI\Mate\Command\DiscoverCommand;
 use Symfony\AI\Mate\Command\InitCommand;
 use Symfony\AI\Mate\Command\ServeCommand;
+use Symfony\AI\Mate\Command\StopCommand;
 use Symfony\AI\Mate\Exception\UnsupportedVersionException;
 use Symfony\Component\Console\Application;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * @author Johannes Wachter <johannes@sulu.io>
@@ -30,7 +35,7 @@ final class App
     public const NAME = 'Symfony AI Mate';
     public const VERSION = '0.1.0';
 
-    public static function build(ContainerBuilder $container): Application
+    public static function build(ContainerInterface $container): Application
     {
         $logger = $container->get(LoggerInterface::class);
         \assert($logger instanceof LoggerInterface);
@@ -44,9 +49,18 @@ final class App
         $application = new Application(self::NAME, self::VERSION);
 
         self::addCommand($application, new InitCommand($rootDir));
-        self::addCommand($application, new ServeCommand($logger, $container));
+        self::addCommand($application, new ServeCommand($container, $logger));
         self::addCommand($application, new DiscoverCommand($rootDir, $logger));
+        self::addCommand($application, new StopCommand((string) $container->getParameter('mate.cache_dir')));
+        self::addCommand($application, new DebugCapabilitiesCommand($logger, $container));
+        self::addCommand($application, new DebugExtensionsCommand($logger, $container));
         self::addCommand($application, new ClearCacheCommand($cacheDir));
+
+        if (\defined('SIGUSR1') && class_exists(RunnerControl::class)) {
+            $application->getSignalRegistry()->register(\SIGUSR1, function () {
+                RunnerControl::$state = RunnerState::STOP;
+            });
+        }
 
         return $application;
     }
@@ -62,7 +76,7 @@ final class App
         } elseif (method_exists($application, 'add')) {
             $application->add($command);
         } else {
-            throw UnsupportedVersionException::forConsole();
+            throw new UnsupportedVersionException('Unsupported version of symfony/console. We cannot add commands.');
         }
     }
 }