+++ src/mate/src/Command/ServeCommand.php
@@ -19,13 +19,13 @@ use Mcp\Server\Session\FileSessionStore;
 use Mcp\Server\Transport\StdioTransport;
 use Psr\Log\LoggerInterface;
 use Symfony\AI\Mate\App;
-use Symfony\AI\Mate\Discovery\ComposerTypeDiscovery;
 use Symfony\AI\Mate\Discovery\FilteredDiscoveryLoader;
-use Symfony\AI\Mate\Discovery\ServiceDiscovery;
+use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
  * Starts the MCP server with stdio transport.
@@ -33,18 +33,38 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
  * @author Johannes Wachter <johannes@sulu.io>
  * @author Tobias Nyholm <tobias.nyholm@gmail.com>
  */
+#[AsCommand('serve', 'Starts the MCP server with stdio transport')]
 class ServeCommand extends Command
 {
-    private ComposerTypeDiscovery $discovery;
+    private string $cacheDir;
+    private FilteredDiscoveryLoader $loader;
 
     public function __construct(
+        private ContainerInterface $container,
         private LoggerInterface $logger,
-        private ContainerBuilder $container,
     ) {
         parent::__construct(self::getDefaultName());
+
         $rootDir = $container->getParameter('mate.root_dir');
         \assert(\is_string($rootDir));
-        $this->discovery = new ComposerTypeDiscovery($rootDir, $logger);
+
+        $cacheDir = $container->getParameter('mate.cache_dir');
+        \assert(\is_string($cacheDir));
+        $this->cacheDir = $cacheDir;
+
+        $extensions = $this->container->getParameter('mate.extensions') ?? [];
+        \assert(\is_array($extensions));
+
+        $disabledFeatures = $this->container->getParameter('mate.disabled_features') ?? [];
+        \assert(\is_array($disabledFeatures));
+
+        $this->loader = new FilteredDiscoveryLoader(
+            basePath: $rootDir,
+            extensions: $extensions,
+            disabledFeatures: $disabledFeatures,
+            discoverer: new Discoverer($logger),
+            logger: $logger
+        );
     }
 
     public static function getDefaultName(): string
@@ -57,34 +77,25 @@ class ServeCommand extends Command
         return 'Starts the MCP server with stdio transport';
     }
 
+    protected function configure(): void
+    {
+        $this->addOption('force-keep-alive', null, InputOption::VALUE_NONE, 'Force a restart of the server if it stops.');
+    }
+
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $rootDir = $this->container->getParameter('mate.root_dir');
-        \assert(\is_string($rootDir));
+        if ($input->getOption('force-keep-alive')) {
+            $output->writeln('The option --force-keep-alive requires using the "bin/mate" file. Try running "./vendor/bin/mate serve --force-keep-alive"');
 
-        $cacheDir = $this->container->getParameter('mate.cache_dir');
-        \assert(\is_string($cacheDir));
-
-        $discovery = new Discoverer($this->logger);
-        $extensions = $this->getExtensionsToLoad();
-        (new ServiceDiscovery())->registerServices($discovery, $this->container, $rootDir, $extensions);
-
-        $disabledVendorFeatures = $this->container->getParameter('mate.disabled_features') ?? [];
-        \assert(\is_array($disabledVendorFeatures));
-        /* @var array<string, array<string, array{enabled: bool}>> $disabledVendorFeatures */
-
-        $this->container->compile();
-
-        $loader = new FilteredDiscoveryLoader(
-            basePath: $rootDir,
-            extensions: $extensions,
-            disabledFeatures: $disabledVendorFeatures,
-            discoverer: $discovery,
-            logger: $this->logger
-        );
+            return Command::INVALID;
+        }
 
         $server = Server::builder()
-            ->setProtocolVersion(ProtocolVersion::tryFrom($this->container->getParameter('mate.mcp_protocol_version') ?? ProtocolVersion::V2025_03_26->value) ?? ProtocolVersion::V2025_03_26)
+            ->setProtocolVersion(
+                ProtocolVersion::tryFrom(
+                    $this->container->getParameter('mate.mcp_protocol_version') ?? ProtocolVersion::V2025_03_26->value
+                ) ?? ProtocolVersion::V2025_03_26
+            )
             ->setServerInfo(
                 App::NAME,
                 App::VERSION,
@@ -93,38 +104,25 @@ class ServeCommand extends Command
                 'https://symfony.com/doc/current/ai/components/mate.html',
             )
             ->setContainer($this->container)
-            ->addLoader($loader)
-            ->setSession(new FileSessionStore($cacheDir.'/sessions'))
+            ->addLoader($this->loader)
+            ->setSession(new FileSessionStore($this->cacheDir.'/sessions'))
             ->setLogger($this->logger)
             ->build();
 
-        $server->run(new StdioTransport());
-
-        return Command::SUCCESS;
-    }
-
-    /**
-     * @return array<string, array{dirs: string[], includes: string[]}>
-     */
-    private function getExtensionsToLoad(): array
-    {
-        $rootDir = $this->container->getParameter('mate.root_dir');
-        \assert(\is_string($rootDir));
-
-        $packageNames = $this->container->getParameter('mate.enabled_extensions');
-        \assert(\is_array($packageNames));
-        /** @var array<int, string> $packageNames */
-
-        /** @var array<string, array{dirs: array<string>, includes: array<string>}> $extensions */
-        $extensions = [];
-
-        foreach ($this->discovery->discover($packageNames) as $packageName => $data) {
-            $extensions[$packageName] = $data;
+        $pidFileName = \sprintf('%s/server_%d.pid', $this->cacheDir, getmypid());
+        if (false === @file_put_contents($pidFileName, (string) getmypid())) {
+            $this->logger->warning('Failed to create PID file', ['path' => $pidFileName]);
         }
 
-        $extensions['_custom'] = $this->discovery->discoverRootProject();
+        try {
+            $server->run(new StdioTransport());
+        } finally {
+            if (file_exists($pidFileName)) {
+                @unlink($pidFileName);
+            }
+        }
 
-        return $extensions;
+        return Command::SUCCESS;
     }
 
     /**